Update PersonneRepository Long --> Integer

+++ src/fr/cdamassy2021/repository/PersonneRepository.java
@@ -7,15 +7,18 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.cdamassy2021.entity.Personne;
 
-public interface PersonneRepository extends CrudRepository<Personne, Long> {
+public interface PersonneRepository extends CrudRepository<Personne, Integer> {
 
 	@Query(value = "SELECT p.id_personne,p.nom,p.prenom, p.email, p.pwd, p.tel, "
-			+ "p.est_formateur, p.est_gestionnaire, p.est_administrateur\n" + "FROM membre_canal mc\n"
-			+ "		INNER JOIN personne p\n" + "			ON mc.id_personne = p.id_personne\n"
+			+ "p.est_formateur, p.est_gestionnaire, p.est_administrateur\n"
+			+ "FROM membre_canal mc\n"
+			+ "		INNER JOIN personne p\n"
+			+ "			ON mc.id_personne = p.id_personne\n"
 			+ "WHERE id_canal=?", nativeQuery = true)
 	public Collection<Personne> findMembreByCanal(int idCanal);
 
-	@Query(value = "DELETE FROM membre_canal WHERE id_personne=? AND id_canal=?", nativeQuery = true)
+	@Query(value = "DELETE FROM membre_canal WHERE id_personne=? AND id_canal=?"
+			, nativeQuery = true)
 	public void SupprimerMembrDuCanal(int idCanal, int idPersonne);
 
 }
